Expand $^O at build time when generating Config::_V()
Config::_V() is how perl -V is implemented internally, and is called called
directly from the interpreter's switch passing routine. Hence the value of
$^O at runtime will always be the same as the value of $^O at build time,
so conditionals dependent on the value of $^O will always take the same
execution path. So only generate the relevant code, and hence avoid shipping
VMS and Cygwin specific code except on those platforms. (But add a sanity
test in Config::_V() to ensure that the runtime $^O has the correct value -
ie that the Perl code and the perl binary correspond.)