EBCDIC has the unicode bug too
We have not had a working modern Perl on EBCDIC for some years.  When I
started out, comments and code led me to conclude erroneously that
natively it supported semantics for all 256 characters 0-255.  It turns
out that I was wrong; it natively (at least on some platforms) has the
same rules (essentially none) for the characters which don't correspond
to ASCII onees, as the rules for these on ASCII platforms.

This commit forces those rules on EBCDIC platforms (even should there be
one that natively uses all 256).  To get all 256, the same things like
'use feature "unicode_strings"' must now be done.